backend/requirements.txt
# Production requirements for Work Management System
# Generated with: pip freeze
psycopg2-binary>=2.9 
# Core Django
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.11

# CORS
django-cors-headers==4.9.0

# Image processing
Pillow==12.0.0

# Static files handling
whitenoise==6.6.0

# Environment variables
python-dotenv==1.1.1

# JWT
PyJWT==2.10.1

# Django dependencies
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
